+++ src/api/snippets/importSnippets.js
@@ -4,22 +4,27 @@ import fs from 'node:fs';
 import JSONStream from 'JSONStream';
 import path from 'node:path';
 
-import { configReader,
+import { addSnippetPwyllCall } from '../pwyllServerCalls.js';
+import {
     errorHandler,
-    infoHandler,
-    checkVersion } from './util.js';
-import { addSnippetPwyllCall } from './pwyllServerCalls.js';
+    PwyllCLIError,
+} from '../../handlers/errorHandler.js';
+import { infoHandler } from '../../handlers/infoHandler.js';
+import { configReader } from '../../handlers/configHandler.js';
+import { checkVersion } from '../../utils/index.js';
 
 // TODO: re-think this...
 export default async function importsToPwyll(file) {
     try {
         const config = await configReader();
         await checkVersion(config);
-        if (!fs.existsSync(file)) {
-            throw new Error(`The file ${file} to import does not exists`);
-        }
         if (path.extname(file) !== '.json') {
-            throw new Error(`The provided file ${file} has not json extension`);
+            throw new PwyllCLIError(
+                `The provided file ${file} has not json extension`);
+        }
+        if (!fs.existsSync(file)) {
+            throw new PwyllCLIError(
+                `The file ${file} to import does not exists`);
         }
         return new Promise((resolve, reject) => {
             const snippets = [];
@@ -35,7 +40,9 @@ export default async function importsToPwyll(file) {
                 .on('end', () => {
                     Promise.all(snippets)
                         .then(() => {
-                            infoHandler(`imported a total of ${snippets.length} snippets into pwyll`);
+                            infoHandler('imported a total of ' +
+                                `${snippets.length}` +
+                                ' snippets into pwyll');
                             return resolve();
                         })
                         .catch((error) => {
@@ -47,7 +54,7 @@ export default async function importsToPwyll(file) {
                 });
         });
     } catch (err) {
-        errorHandler(err);
+        return errorHandler(err);
     }
 }