+++ src/handlers/warningHandler.js
@@ -1,0 +1,6 @@
+/* eslint no-console: "off" */
+import chalk from 'chalk';
+
+export function warningHandler(warningMessage) {
+    console.log(`[${chalk.yellow('WARN')}] ${warningMessage}`);
+}